extracted dialogs from main widget uploader form
